Refactored.

+++ e5/services/fileService.js
@@ -1,39 +1,48 @@
 let fs = require("fs");
+const devLogger = require("../log/devLogger");
 
-const createFile = (path) => {
+const createFile = (path, req, res) => {
   return new Promise((resolve, reject) => {
    try {
       fs.writeFileSync(path,"[]", "utf-8");
       resolve();
     }
     catch(err) {
-      // console.log("file not read error", err);
+      devLogger.error(`${err.message} path: ${req.method} ${req.originalUrl}`);
+      res.status(500);
       reject("File cannot be created");
     }
   });
 };
 
-const readFile = (path) => {
+const readFile = (path,req, res) => {
   return new Promise((resolve, reject) => {
     try {
       let data = fs.readFileSync(path);
       resolve(JSON.parse(data));
     }
     catch(err) {
-      // console.log();
-      reject("File not Found")
+      devLogger.error(`${err.message} path: ${req.method} ${req.originalUrl}`);
+      res.status(500);
+      reject("File not Found");
     }
   });
 };
 
-const writeFile = (path, data) => {
+const writeFile = (path, data, req, res) => {
   return new Promise((resolve, reject) => {
     try {
+      if(!fs.existsSync(path)) {
+        devLogger.warn(`File doesnot exists path: ${req.method} ${req.originalUrl}`);
+        reject("File doesnot exists");
+        return;
+      }
       fs.writeFileSync(path, JSON.stringify(data));
       resolve();
     }
     catch(err) {
-      // console.log(err);
+      devLogger.error(`${err.message} path: ${req.method} ${req.originalUrl}`);
+      res.status(500);
       reject("Cannot write in file");
     }
   });
